requirements.txt
Django==2.2
djangorestframework==3.9.2
requests==2.21
PyGreSQL==5.0.7
psycopg2==2.8.2
pyjwt